Moving the udp unit tests to UdpIT
