+++ driver/src/test/resources/org/kaazing/k3po/driver/internal/udp/udp.server.rpt
@@ -1,0 +1,45 @@
+#
+# Copyright 2007-2015, Kaazing Corporation. All rights reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+
+accept udp://localhost:8080
+       option timeout 1000
+accepted
+connected
+read "Hello11"
+read "Hello12"
+write "Hello World11"
+write "Hello World12"
+# closed because of idle timeout
+closed
+
+accepted
+connected
+read "Hello21"
+read "Hello22"
+write "Hello World21"
+write "Hello World22"
+close
+closed
+
+accept udp://localhost:8081
+accepted
+connected
+read "Hello31"
+read "Hello32"
+write "Hello World31"
+write "Hello World32"
+close
+closed
